Add command line arguments for i/o filenames
I was iniitally attempting to concatenate each input file's
output (containing sumed and sorted frequencies) to the
chosen CSV output file. This was not ideal because deleting CSV rows
in-line requires a file pointer to scan the whole file multiple times
and gets very confusing since you want to use two file pointers at a
given time and have to reset your pointer location each time.

I am going to instead try to complete all of the suming
and sorting processing with the data as a list before
writing the entire processed list out to a CSV file.
My data set is not very large and apparently passing entire lists as an
argument to a function is not very inefficient in python. The internet suggests
that the variable is not actually passed by value when you do this.