Optimized CWE frequency count by presorting CWEs
The previous attempt at an algorithm was going to be too messy.
This method not only actually works this time but is simpler because
I do not have to modify the raw source list of CWE's where there
are as many non-unique CWE lines as the frequency that a CWE-ID occured.
I simply compare a temporary variable to the value of the next line in
the source file and increment a counter each time a see a duplicate
line. When a line differs from the temporary variable value, you know
that you are reading a different word, so you may reset and restart the
incrementing of your frequency counter.