Add CSV Name column display via stdout
It appears as though you cannot actually write to the same file that you
are reading. I need to merge this getTitles.py with the countCWE.py
anyway, so that should allow me to write to the file when it is first
created instead of trying to edit an already created file.