Add names to CWE listings when available
Requires a dictionary of sorts as provided by mitre.
The cli argument to include a name/title is optional.
But the file used as a dictionary must have the format where the first
column has ID's with only the number portion not the 'CWE-' part and the
second column has to have the name of the CWE.